requirements.txt
torch
torchvision
pillow
tensorboard
tqdm
numpy
matplotlib
